move LfSelectable to shared instead of internal-shared so it is exposed in npm package
