+++ projects/ui-components/lf-repository-browser/lf-repository-browser.component.ts
@@ -4,7 +4,8 @@ import { LfTreeNodeService, LfTreeNode } from './ILfTreeNodeService';
 import { Subject } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
 import { LfSelectionListComponent, SelectedItemEvent } from './../lf-selection-list/lf-selection-list-public-api';
-import { AppLocalizationService, ILfSelectable, ItemWithId } from '../internal-shared/internal-shared-public-api';
+import { AppLocalizationService } from '../internal-shared/internal-shared-public-api';
+import { ILfSelectable, ItemWithId } from '../shared/lf-shared-public-api';
 
 @Component({
   selector: 'lf-repository-browser-component',
