+++ projects/ui-components/shared/lf-shared-public-api.ts
@@ -1,3 +1,4 @@
 export * from './enums';
 export * from './lf-breadcrumbs/lf-breadcrumbs-public-api';
 export * from './lf-toolbar/lf-toolbar-public-api';
+export * from './LfSelectable';
